+++ services/console/src/components/console/billing/plan/Plan.tsx
@@ -1,16 +1,17 @@
 import type { InitOutput } from "bencher_valid";
-import { createMemo, createResource, For, Resource } from "solid-js";
+import { createMemo, createResource, For, type Resource } from "solid-js";
 import {
 	CardBrand,
-	JsonAuthUser,
-	JsonPlan,
-	JsonUsage,
+	type JsonAuthUser,
+	type JsonPlan,
+	type JsonUsage,
 	PlanLevel,
 	PlanStatus,
 } from "../../../../types/bencher";
 import type { Params } from "astro";
 import { validJwt } from "../../../../util/valid";
 import { httpGet } from "../../../../util/http";
+import { dateTimeMillis } from "../../../../util/convert";
 
 const fmtDateTime = (date_str: undefined | string) => {
 	if (date_str === undefined) {
@@ -21,20 +22,6 @@ const fmtDateTime = (date_str: undefined | string) => {
 		const date = new Date(date_ms);
 		if (date) {
 			return date.toDateString();
-		}
-	}
-	return null;
-};
-
-const dateTimeMillis = (date_str: undefined | string) => {
-	if (date_str === undefined) {
-		return null;
-	}
-	const date_ms = Date.parse(date_str);
-	if (date_ms) {
-		const date = new Date(date_ms);
-		if (date) {
-			return date.getTime();
 		}
 	}
 	return null;