+++ changelog.md
@@ -7,6 +7,8 @@ sortOrder: 4
 - Add ability for `bencher run` to read output from a file without running a command argument
 - Improve Report creation error messages
 - Add filtering Reports by Branch, Testbed, and start/end time
+- Add one month historical default when viewing Reports in the Perf Plot
+- Remove old UI source and scripts
 
 ## `v0.3.12`
 - Change Metric `lower_bound` and `upper_bound` to `lower_value` and `upper_value` respectively