+++ changelog.md
@@ -3,7 +3,8 @@ title: "Changelog"
 heading: "Bencher Changelog"
 sortOrder: 4
 ---
-## Pending `v0.3.13`
+
+## `v0.3.13`
 - Add ability for `bencher run` to read output from a file without running a command argument
 - Improve Report creation error messages
 - Add filtering Reports by Branch, Testbed, and start/end time